data-generator/pyproject.toml
[tool.poetry]
name = "grid-data-generator"
version = "0.1.0"
description = "Smart grid sensor data generator for anomaly detection pipeline"
authors = ["Grid Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
requests = "^2.31.0"
kafka-python = "^2.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate = "generator:main"
